Move and rename deepLink APIs
